requirements.txt
python-telegram-bot==13.15
Flask==2.1.1
requests==2.26.0
pybit==1.1.0
openai==0.27.0
cryptography==3.4.8
python-dotenv==0.19.0
psutil==5.8.0
gitpython==3.1.24

